requirements.txt
pygithub==1.37
